skills/create-storyboard/pyproject.toml
[project]
name = "create-storyboard"
version = "0.1.0"
description = "Transform screenplay to animatic with camera framing"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pillow>=10.0.0",
]

[project.scripts]
create-storyboard = "orchestrator:app"
